Split asset code out of Projects file
